feat: Add transform and clean a little
