+++ src/evo_lib/types/quaternion.py
@@ -1,0 +1,1 @@
+# TODO: Move quaternion code from pose.py
